added gpt 4o mini

+++ my_gemini.py
@@ -254,7 +254,7 @@ def chat(query: str,
 
 
 @cachetools.func.ttl_cache(maxsize=10, ttl=10 * 60)
-def img2txt(data_: bytes, prompt: str = "Что на картинке, подробно?", temp: int = 1, model: str = 'gemini-1.5-flash') -> str:
+def img2txt(data_: bytes, prompt: str = "Что на картинке, подробно?", temp: float = 1, model: str = 'gemini-1.5-flash') -> str:
     '''Convert image to text.
     '''
     try:
